# pkg/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Message:
    properties:
      message:
        description: A standard message.
        type: string
    type: object
  models.Permission:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - PermissionRead
    - PermissionWrite
    - PermissionAdmin
  models.Project:
    properties:
      background_blur_hash:
        description: Contains a very small version of the project background to use
          as a blurry preview until the actual background is loaded. Check out https://blurha.sh/
          to learn how it works.
        type: string
      background_information:
        description: Holds extra information about the background set since some background
          providers require attribution or similar. If not null, the background can
          be accessed at /projects/{projectID}/background
      created:
        description: A timestamp when this project was created. You cannot change
          this value.
        type: string
      description:
        description: The description of the project.
        type: string
      hex_color:
        description: The hex color of this project
        maxLength: 7
        type: string
      id:
        description: The unique, numeric id of this project.
        type: integer
      identifier:
        description: The unique project short identifier. Used to build task identifiers.
        maxLength: 10
        minLength: 0
        type: string
      is_archived:
        description: Whether a project is archived.
        type: boolean
      is_favorite:
        description: True if a project is a favorite. Favorite projects show up in
          a separate parent project. This value depends on the user making the call
          to the api.
        type: boolean
      max_permission:
        $ref: '#/definitions/models.Permission'
      owner:
        allOf:
        - $ref: '#/definitions/user.User'
        description: The user who created this project.
      parent_project_id:
        type: integer
      position:
        description: The position this project has when querying all projects. See
          the tasks.position property on how to use this.
        type: number
      subscription:
        allOf:
        - $ref: '#/definitions/models.Subscription'
        description: |-
          The subscription status for the user reading this project. You can only read this property, use the subscription endpoints to modify it.
          Will only returned when retreiving one project.
      title:
        description: The title of the project. You'll see this in the overview.
        maxLength: 250
        minLength: 1
        type: string
      updated:
        description: A timestamp when this project was last updated. You cannot change
          this value.
        type: string
      views:
        items:
          $ref: '#/definitions/models.ProjectView'
        type: array
    type: object
...
  models.ProjectView:
    properties:
      bucket_configuration:
        description: When the bucket configuration mode is not `manual`, this field
          holds the options of that configuration.
        items:
          $ref: '#/definitions/models.ProjectViewBucketConfiguration'
        type: array
      bucket_configuration_mode:
        description: The bucket configuration mode. Can be `none`, `manual` or `filter`.
          `manual` allows to move tasks between buckets as you normally would. `filter`
          creates buckets based on a filter for each bucket.
        enum:
        - none
        - manual
        - filter
        - manual
        type: string
      created:
        description: A timestamp when this reaction was created. You cannot change
          this value.
        type: string
      default_bucket_id:
        description: The ID of the bucket where new tasks without a bucket are added
          to. By default, this is the leftmost bucket in a view.
        type: integer
      done_bucket_id:
        description: If tasks are moved to the done bucket, they are marked as done.
          If they are marked as done individually, they are moved into the done bucket.
        type: integer
      filter:
        allOf:
        - $ref: '#/definitions/models.TaskCollection'
        description: The filter query to match tasks by. Check out https://vikunja.io/docs/filters
          for a full explanation.
      id:
        description: The unique numeric id of this view
        type: integer
      position:
        description: The position of this view in the list. The list of all views
          will be sorted by this parameter.
        type: number
      project_id:
        description: The project this view belongs to
        type: integer
      title:
        description: The title of this view
        type: string
      updated:
        description: A timestamp when this view was updated. You cannot change this
          value.
        type: string
      view_kind:
        description: The kind of this view. Can be `list`, `gantt`, `table` or `kanban`.
        enum:
        - list
        - gantt
        - table
        - kanban
        type: string
    type: object
  models.ProjectViewBucketConfiguration:
    properties:
      filter:
        $ref: '#/definitions/models.TaskCollection'
      title:
        type: string
    type: object
...
  models.Subscription:
    properties:
      created:
        description: A timestamp when this subscription was created. You cannot change
          this value.
        type: string
      entity:
        type: integer
      entity_id:
        description: The id of the entity to subscribe to.
        type: integer
      id:
        description: The numeric ID of the subscription
        type: integer
    type: object
...
  models.TaskCollection:
    properties:
      filter:
        description: The filter query to match tasks by. Check out https://vikunja.io/docs/filters
          for a full explanation.
        type: string
      filter_include_nulls:
        description: If set to true, the result will also include null values
        type: boolean
      order_by:
        description: The query parameter to order the items by. This can be either
          asc or desc, with asc being the default.
        items:
          type: string
        type: array
      s:
        type: string
      sort_by:
        description: The query parameter to sort by. This is for ex. done, priority,
          etc.
        items:
          type: string
        type: array
    type: object
...
  user.User:
    properties:
      created:
        description: A timestamp when this task was created. You cannot change this
          value.
        type: string
      email:
        description: The user's email address.
        maxLength: 250
        type: string
      id:
        description: The unique, numeric id of this user.
        type: integer
      name:
        description: The full name of the user.
        type: string
      updated:
        description: A timestamp when this task was last updated. You cannot change
          this value.
        type: string
      username:
        description: The username of the user. Is always unique.
        maxLength: 250
        minLength: 1
        type: string
    type: object
...
info:
  contact:
    email: hello@vikunja.io
    name: General Vikunja contact
    url: https://vikunja.io/contact/
  description: |-
    # Pagination
    Every endpoint capable of pagination will return two headers:
    * `x-pagination-total-pages`: The total number of available pages for this request
    * `x-pagination-result-count`: The number of items returned for this request.
    # Permissions
    All endpoints which return a single item (project, task, etc.) - no array - will also return a `x-max-permission` header with the max permission the user has on this item as an int where `0` is `Read Only`, `1` is `Read & Write` and `2` is `Admin`.
    This can be used to show or hide ui elements based on the permissions the user has.
    # Errors
    All errors have an error code and a human-readable error message in addition to the http status code. You should always check for the status code in the response, not only the http status code.
    Due to limitations in the swagger library we're using for this document, only one error per http status code is documented here. Make sure to check the [error docs](https://vikunja.io/docs/errors/) in Vikunja's documentation for a full list of available error codes.
    # Authorization
    **JWT-Auth:** Main authorization method, used for most of the requests. Needs `Authorization: Bearer <jwt-token>`-header to authenticate successfully.

...

    **BasicAuth:** Only used when requesting tasks via CalDAV.
...
  license:
...
paths:
...
  /{username}/avatar:
    get:
      description: Returns the user avatar as image.
      parameters:
      - description: The username of the user who's avatar you want to get
        in: path
        name: username
        required: true
        type: string
      - description: The size of the avatar you want to get. If bigger than the max
          configured size this will be adjusted to the maximum size.
        in: query
        name: size
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: The avatar
          schema:
            type: file
        "404":
          description: The user does not exist.
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Message'
      summary: User Avatar
      tags:
      - user
...
  /projects/{id}/background:
    delete:
      description: Removes a previously set project background, regardless of the
        project provider used to set the background. It does not throw an error if
        the project does not have a background.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: The project
          schema:
            $ref: '#/definitions/models.Project'
        "403":
          description: No access to this project.
          schema:
            $ref: '#/definitions/models.Message'
        "404":
          description: The project does not exist.
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Remove a project background
      tags:
      - project
    get:
      description: Get the project background of a specific project. **Returns json
        on error.**
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: The project background file.
          schema:
            type: file
        "403":
          description: No access to this project.
          schema:
            $ref: '#/definitions/models.Message'
        "404":
          description: The project does not exist.
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Get the project background
      tags:
      - project
...
  /user/settings/avatar/upload:
    put:
      consumes:
      - multipart/form-data
      description: Upload a user avatar. This will also set the user's avatar provider
        to "upload"
      parameters:
      - description: The avatar as single file.
        in: formData
        name: avatar
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The avatar was set successfully.
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: File is no image.
          schema:
            $ref: '#/definitions/models.Message'
        "403":
          description: File too large.
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Upload a user avatar
      tags:
      - user
...
securityDefinitions:
  BasicAuth:
    type: basic
  JWTKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"